+++ ServerEB/EscapeBandages/src/main/java/es/sidelab/EscapeBandages/UsersManager.java
@@ -30,7 +30,7 @@ public class UsersManager {
 			Stack<Client> clientsToDisconnect = new Stack<>();
 			
 			
-			for(Client client: UsersController.clients()) {
+			for(Client client: ClientController.clients()) {
 				//System.out.println("UwU");
 				client.increaseInactivty();
 				if(client.getTimeOfInactivity() > maxTimeOfInactivity) {
